chore: rename ProtectedRoute to Auth
