+++ packages/round-manager/src/features/round/ViewApplicationPage.tsx
@@ -8,7 +8,7 @@ import {
 import { useListRoundsQuery } from "../api/services/round"
 import ConfirmationModal from "../common/ConfirmationModal"
 import Navbar from "../common/Navbar"
-import { useWallet } from "../common/ProtectedRoute"
+import { useWallet } from "../common/Auth"
 import { Button } from "../common/styles"
 import { ReactComponent as TwitterIcon } from "../../assets/twitter-logo.svg"
 import { ReactComponent as GithubIcon } from "../../assets/github-logo.svg"
